Moved definitions of allow_string_paths() and allow_pathlib_paths() to fileutils.pathutils (keeping reference for namespace backwards compat)
