+++ polymerist/genutils/fileutils/pathutils.py
@@ -3,18 +3,69 @@
 __author__ = 'Timotej Bernat'
 __email__ = 'timotej.bernat@colorado.edu'
 
-from typing import Union
+from typing import Callable, Concatenate, ParamSpec, TypeVar, Union
+T = TypeVar('T')
+Params = ParamSpec('Params')
+
+from inspect import signature
+from functools import wraps
+ 
 from pathlib import Path
+
+from ..decorators.meta import extend_to_methods
+from ..decorators.signatures import modify_param_annotation_by_index
     
 
-# PATH PROPERTY FUNCTIONS (DON'T MODIFY ANYTHING)
-def _dotless(extension : str) -> str:
-    '''Separate the dot from a SINGLE extension file suffix. Returns the original suffix if not dot is present'''
-    return extension.split('.')[-1] 
+# PATH CONVERSION FUNCTIONS (FOR CHANGING BETWEEN TYPES)
+def aspath(path : Union[str, Path]) -> Path:
+	'''Allow functions which expect Paths to also accept strings'''
+	if not isinstance(path, Path):
+		path = Path(path)
+	return path
 
-def dotless(path : Path) -> str:
-    '''Separate the dot from file path. Returns the original suffix if not dot is present'''
-    return _dotless(path.suffix)
+def asstrpath(strpath : Union[str, Path]) -> str:
+	'''Allow functions which expect strings paths to also accept Paths'''
+	if not isinstance(strpath, str):
+		strpath = str(strpath)
+	return strpath
+
+@extend_to_methods
+def allow_string_paths(funct : Callable[[Concatenate[Path, Params]], T]) -> Callable[[Concatenate[Union[Path, str], Params]], T]:
+    '''Modifies a function which expects a Path as its first argument to also accept string-paths'''
+    # TODO : add assertion that the wrapped function has at least one arg AND that the first arg is of the desired (limited) type
+    old_sig = signature(funct) # lookup old type signature
+
+    @wraps(funct) # for preserving docstring and type annotations / signatures
+    def str_path_wrapper(flex_path : Union[str, Path], *args : Params.args, **kwargs : Params.kwargs) -> T:
+        '''First converts stringy paths into normal Paths, then executes the original function'''
+        return funct(aspath(flex_path), *args, **kwargs)
+
+    # MODIFY SIGNATURE OF PATH-LIKE FIRST ARGUMENT TO MATCH NEW TYPE FLEXIBILITY
+    str_path_wrapper.__signature__ = modify_param_annotation_by_index(
+        old_sig,
+        index=0, # modify signature of first argument to reflect new type flexibility
+        new_type=Union[Path, str]
+    )
+    return str_path_wrapper
+
+@extend_to_methods
+def allow_pathlib_paths(funct : Callable[[Concatenate[str, Params]], T]) -> Callable[[Concatenate[Union[Path, str], Params]], T]:
+    '''Modifies a function which expects a string path as its first argument to also accept canonical pathlib Paths'''
+    # TODO : add assertion that the wrapped function has at least one arg AND that the first arg is of the desired (limited) type
+    old_sig = signature(funct) # lookup old type signature
+
+    @wraps(funct) # for preserving docstring and type annotations / signatures
+    def str_path_wrapper(flex_path : Union[str, Path], *args : Params.args, **kwargs : Params.kwargs) -> T:
+        '''First converts normal Paths into stringy paths, then executes the original function'''
+        return funct(asstrpath(flex_path), *args, **kwargs)
+
+    # MODIFY SIGNATURE OF PATH-LIKE FIRST ARGUMENT TO MATCH NEW TYPE FLEXIBILITY
+    str_path_wrapper.__signature__ = modify_param_annotation_by_index(
+        old_sig,
+        index=0, # modify signature of first argument to reflect new type flexibility
+        new_type=Union[Path, str]
+    )
+    return str_path_wrapper
 
 
 # EMPTINESS CHECKS
@@ -22,7 +73,7 @@ def is_empty_dir(dirpath : Path) -> bool:
     '''Check if a directory contains no files'''
     if not dirpath.is_dir():
          raise NotADirectoryError(f'dirpath must point to directory, not to file "{dirpath}"')
-    return list(dirpath.iterdir()) == [] # can't use "len" for generators : TODO : make this more efficient (i.e. iteration-based) for large directories
+    return any(dirpath.iterdir()) # can't use "len" for generators : TODO : make this more efficient (i.e. iteration-based) for large directories
 
 def is_empty_file(filepath : Path) -> bool:
     '''Check if a file contains no data'''
@@ -42,9 +93,8 @@ def is_empty(path : Path) -> bool:
     elif path.is_file():
         return is_empty_file(path)
     else:
-        return FileNotFoundError
+        raise FileNotFoundError(f'Path "{path}" does not exist')
     
-
 # PATH CREATION FUNCTIONS
 def assemble_path(directory : Path, prefix : str, extension : str, postfix : str='') -> Path:
     '''Combine output, naming, descriptive, and filetype info to generate a complete Path'''
@@ -54,20 +104,14 @@ def assemble_path(directory : Path, prefix : str, extension : str, postfix : str
 
     return directory / path_name
 
+# PATH PROPERTY FUNCTIONS (DON'T MODIFY ANYTHING)
+def _dotless(extension : str) -> str:
+    '''Separate the dot from a SINGLE extension file suffix. Returns the original suffix if not dot is present'''
+    return extension.split('.')[-1] 
 
-# PATH CONVERSION FUNCTIONS (FOR CHANGING BETWEEN TYPES)
-def aspath(path : Union[str, Path]) -> Path:
-	'''Allow functions which expect Paths to also accept strings'''
-	if not isinstance(path, Path):
-		path = Path(path)
-	return path
-
-def asstrpath(strpath : Union[str, Path]) -> str:
-	'''Allow functions which expect strings paths to also accept Paths'''
-	if not isinstance(strpath, str):
-		strpath = str(strpath)
-	return strpath
-
+def dotless(path : Path) -> str:
+    '''Separate the dot from file path. Returns the original suffix if not dot is present'''
+    return _dotless(path.suffix)
 
 # PATH MODIFICATION FUNCTIONS (CHANGING THE STRUCTURE OF A PATH OBJECT)
 def default_suffix(path : Path, suffix : str) -> Path:
